Split TS generation into files : move options handling to separate source files
