+++ src/tools/ts-generator/linksTSgenerator.h
@@ -1,0 +1,11 @@
+
+#pragma once
+
+class LinksTSgenerator
+{
+    void loadData();
+    void generate();
+
+
+};
+
